+++ src/cucumber.rs
@@ -949,6 +949,22 @@ where
         self
     }
 
+    /// Makes stop running tests on the first failure.
+    ///
+    /// __NOTE__: All the already started [`Scenario`]s at the moment of failure
+    ///           will be finished.
+    ///
+    /// __NOTE__: Retried [`Scenario`]s are considered as failed, only in case
+    ///           they exhaust all retry attempts and still do fail.
+    ///
+    /// [`Scenario`]: gherkin::Scenario
+    #[allow(clippy::missing_const_for_fn)] // false positive: drop in const
+    #[must_use]
+    pub fn fail_fast(mut self) -> Self {
+        self.runner = self.runner.fail_fast();
+        self
+    }
+
     /// Makes failed [`Scenario`]s being retried after the specified
     /// [`Duration`] passes.
     ///